+++ lib/root_page.dart
@@ -10,12 +10,13 @@ import 'package:chronopsi/pages/connection_page.dart';
 
 import 'pages/home_page.dart';
 
-enum AppState {loading, connected, disconnected }
+enum AppState { loading, connected, disconnected }
 
 class RootPage extends StatefulWidget {
   RootPage({Key key, @required this.adaptiveThemeMode}) : super(key: key);
 
   final AdaptiveThemeMode adaptiveThemeMode;
+
   @override
   _RootPageState createState() => _RootPageState();
 }
@@ -39,8 +40,9 @@ class _RootPageState extends State<RootPage> {
     //init Moor Database
     configuration.localMoorDatabase = LocalMoorDatabase();
     //get packageInfo
-    configuration.packageInfo =
-        Platform.isWindows || Platform.isLinux ? null : await PackageInfo.fromPlatform();
+    configuration.packageInfo = Platform.isWindows || Platform.isLinux
+        ? null
+        : await PackageInfo.fromPlatform();
 
     //init sharedPreferences
     configuration.sharedPreferences = await SharedPreferences.getInstance();
@@ -56,6 +58,10 @@ class _RootPageState extends State<RootPage> {
             CACHE_KEEPING_DURATION_DEFAULT_VALUE;
 
     configuration.isDarkTheme = widget.adaptiveThemeMode.isDark;
+
+    configuration.doNotShowErrorMsgAgain =
+        configuration.sharedPreferences.getBool(doNotShowErrorMsgAgainKey) ??
+            DO_NOT_SHOW_ERROR_MSG_AGAIN_DEFAULT_VALUE;
 
     if (configuration.sharedPreferences.getString(logInKey) == null ||
         configuration.sharedPreferences.getString(logInKey) == "") {
@@ -77,27 +83,35 @@ class _RootPageState extends State<RootPage> {
     switch (appState) {
       case AppState.loading:
         {
-          return Scaffold(
-            backgroundColor: Color(0xFF2a3035),
-            body: Column(
-              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-              children: [
-                Image.asset('assets/appLogo/logo_avec_padding.png'),
-                CircularProgressIndicator(),
-              ],
+          return WillPopScope(
+            onWillPop: () async => false,
+            child: Scaffold(
+              backgroundColor: Color(0xFF2a3035),
+              body: Column(
+                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                children: [
+                  Image.asset('assets/appLogo/logo_avec_padding.png'),
+                  CircularProgressIndicator(),
+                ],
+              ),
             ),
           );
         }
         break;
       case AppState.connected:
         {
-          return HomePage(configuration: configuration);
+          return WillPopScope(
+              onWillPop: () async => false,
+              child: HomePage(configuration: configuration));
         }
         break;
       case AppState.disconnected:
         {
-          return LogInPage(
-            configuration: configuration,
+          return WillPopScope(
+            onWillPop: () async => false,
+            child: LogInPage(
+              configuration: configuration,
+            ),
           );
         }
         break;